# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ClassLevel = typeof ClassLevel[keyof typeof ClassLevel];


export const ClassLevel = {
  pemula: 'pemula',
  menengah: 'menengah',
  lanjutan: 'lanjutan',
} as const;

export interface InstructorSummary {
  id: string;
  name: string;
  bio: string;
  photoUrl: string;
  classCount: number;
}

export interface ClassInstructor {
  id: string;
  name: string;
  photoUrl: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ClassModule {
  id: string;
  title: string;
  orderIndex: number;
  /** @nullable */
  durationMinutes: number | null;
}

export type ClassDetailStatus = typeof ClassDetailStatus[keyof typeof ClassDetailStatus];


export const ClassDetailStatus = {
  draft: 'draft',
  published: 'published',
} as const;

/**
 * @nullable
 */
export type ClassDetailLevel = typeof ClassDetailLevel[keyof typeof ClassDetailLevel] | null;


export const ClassDetailLevel = {
  pemula: 'pemula',
  menengah: 'menengah',
  lanjutan: 'lanjutan',
} as const;

export interface ClassDetail {
  id: string;
  title: string;
  description: string;
  coverImage: string;
  basePrice: number;
  /** @nullable */
  discountPrice: number | null;
  status: ClassDetailStatus;
  /** @nullable */
  level: ClassDetailLevel;
  /** @nullable */
  category: string | null;
  /** @nullable */
  youtubePlaylistId: string | null;
  /** @nullable */
  gdriveMateriUrl: string | null;
  /** @nullable */
  waGroupUrl: string | null;
  instructor: InstructorSummary;
  modules: ClassModule[];
  moduleCount: number;
  /** @nullable */
  totalDurationMinutes: number | null;
}

export type ClassSummaryStatus = typeof ClassSummaryStatus[keyof typeof ClassSummaryStatus];


export const ClassSummaryStatus = {
  draft: 'draft',
  published: 'published',
} as const;

/**
 * @nullable
 */
export type ClassSummaryLevel = typeof ClassSummaryLevel[keyof typeof ClassSummaryLevel] | null;


export const ClassSummaryLevel = {
  pemula: 'pemula',
  menengah: 'menengah',
  lanjutan: 'lanjutan',
} as const;

export interface ClassSummary {
  id: string;
  title: string;
  description: string;
  coverImage: string;
  basePrice: number;
  /** @nullable */
  discountPrice: number | null;
  status: ClassSummaryStatus;
  /** @nullable */
  level: ClassSummaryLevel;
  /** @nullable */
  category: string | null;
  /** @nullable */
  youtubePlaylistId: string | null;
  instructor: ClassInstructor;
  moduleCount: number;
  /** @nullable */
  totalDurationMinutes: number | null;
}

export type CreateClassRequestStatus = typeof CreateClassRequestStatus[keyof typeof CreateClassRequestStatus];


export const CreateClassRequestStatus = {
  draft: 'draft',
  published: 'published',
} as const;

/**
 * @nullable
 */
export type CreateClassRequestLevel = typeof CreateClassRequestLevel[keyof typeof CreateClassRequestLevel] | null;


export const CreateClassRequestLevel = {
  pemula: 'pemula',
  menengah: 'menengah',
  lanjutan: 'lanjutan',
} as const;

export interface CreateClassRequest {
  title: string;
  description?: string;
  coverImage?: string;
  basePrice?: number;
  /** @nullable */
  discountPrice?: number | null;
  status?: CreateClassRequestStatus;
  /** @nullable */
  level?: CreateClassRequestLevel;
  /** @nullable */
  category?: string | null;
  instructorId: string;
}

export type UpdateClassRequestStatus = typeof UpdateClassRequestStatus[keyof typeof UpdateClassRequestStatus];


export const UpdateClassRequestStatus = {
  draft: 'draft',
  published: 'published',
} as const;

/**
 * @nullable
 */
export type UpdateClassRequestLevel = typeof UpdateClassRequestLevel[keyof typeof UpdateClassRequestLevel] | null;


export const UpdateClassRequestLevel = {
  pemula: 'pemula',
  menengah: 'menengah',
  lanjutan: 'lanjutan',
} as const;

export interface UpdateClassRequest {
  title?: string;
  description?: string;
  coverImage?: string;
  basePrice?: number;
  /** @nullable */
  discountPrice?: number | null;
  status?: UpdateClassRequestStatus;
  /** @nullable */
  level?: UpdateClassRequestLevel;
  /** @nullable */
  category?: string | null;
  instructorId?: string;
}

export interface CreateInstructorRequest {
  name: string;
  bio?: string;
  photoUrl?: string;
}

export interface UpdateInstructorRequest {
  name?: string;
  bio?: string;
  photoUrl?: string;
}

export interface CartItem {
  id: string;
  classId: string;
  addedAt: string;
  class: ClassSummary;
}

export interface AddCartItemRequest {
  userId: string;
  classId: string;
}

export interface CreateCheckoutRequest {
  userId: string;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
} as const;

export interface InvoiceItem {
  id: string;
  classId: string;
  price: number;
  class: ClassSummary;
}

export interface Invoice {
  id: string;
  userId: string;
  totalAmount: number;
  status: InvoiceStatus;
  createdAt: string;
  /** @nullable */
  paidAt: string | null;
  /**
     * URL halaman pembayaran Mayar. Diisi saat Mayar sudah diintegrasikan; null selama mode simulasi.
     * @nullable
     */
  paymentUrl?: string | null;
  items: InvoiceItem[];
}

export interface Testimonial {
  id: string;
  name: string;
  /** @nullable */
  role: string | null;
  content: string;
  photoUrl: string;
  isPublished: boolean;
  orderIndex: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTestimonialRequest {
  name: string;
  /** @nullable */
  role?: string | null;
  content: string;
  photoUrl?: string;
  isPublished?: boolean;
  orderIndex?: number;
}

export interface UpdateTestimonialRequest {
  name?: string;
  /** @nullable */
  role?: string | null;
  content?: string;
  photoUrl?: string;
  isPublished?: boolean;
  orderIndex?: number;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  tagline: string;
  logoUrl: string;
  contactEmail: string;
  contactPhone: string;
  address: string;
  founderName: string;
  founderBio: string;
  founderPhotoUrl: string;
  socialInstagram: string;
  socialYoutube: string;
  socialFacebook: string;
  socialTiktok: string;
  studentCountLabel: string;
  updatedAt: string;
}

export interface UpdateSiteSettingsRequest {
  siteName?: string;
  tagline?: string;
  logoUrl?: string;
  contactEmail?: string;
  contactPhone?: string;
  address?: string;
  founderName?: string;
  founderBio?: string;
  founderPhotoUrl?: string;
  socialInstagram?: string;
  socialYoutube?: string;
  socialFacebook?: string;
  socialTiktok?: string;
  studentCountLabel?: string;
}

export type ListClassesParams = {
search?: string;
level?: ClassLevel;
category?: string;
instructorId?: string;
sort?: ListClassesSort;
/**
 * When true, returns classes of any status (for admin use). Defaults to published only.
 */
includeAll?: boolean;
};

export type ListClassesSort = typeof ListClassesSort[keyof typeof ListClassesSort];


export const ListClassesSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  popular: 'popular',
} as const;

export type ListRecommendedClassesParams = {
userId: string;
limit?: number;
};

export type ListCartItemsParams = {
userId: string;
};