# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout:
    post:
      operationId: createCheckout
      tags: [checkout]
      summary: Create an invoice from the current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutRequest"
      responses:
        "200":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Cart is empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/{id}:
    get:
      operationId: getCheckout
      tags: [checkout]
      summary: Get invoice status (untuk polling setelah redirect dari payment gateway)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/{id}/simulate-success:
    post:
      operationId: simulateCheckoutSuccess
      tags: [checkout]
      summary: Simulate a successful payment webhook for an invoice
      description: Marks the invoice as paid, creates enrollments for each invoice item, and clears the user's cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Invoice marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/{id}/simulate-fail:
    post:
      operationId: simulateCheckoutFail
      tags: [checkout]
      summary: Simulate a failed payment webhook for an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Invoice marked as failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ClassInstructor:
      type: object
      required: [id, name, photoUrl]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        photoUrl:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ClassSummary:
      type: object
      required:
        - id
        - title
        - description
        - coverImage
        - basePrice
        - discountPrice
        - status
        - level
        - category
        - instructor
        - moduleCount
        - totalDurationMinutes
        - youtubePlaylistId
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        basePrice:
          type: integer
        discountPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published]
        level:
          type: ["string", "null"]
          enum: [pemula, menengah, lanjutan, null]
        category:
          type: ["string", "null"]
        youtubePlaylistId:
          type: ["string", "null"]
        instructor:
          $ref: "#/components/schemas/ClassInstructor"
        moduleCount:
          type: integer
        totalDurationMinutes:
          type: ["integer", "null"]
...
    CreateCheckoutRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
    InvoiceStatus:
      type: string
      enum: [pending, paid, failed]
    InvoiceItem:
      type: object
      required: [id, classId, price, class]
      properties:
        id:
          type: string
          format: uuid
        classId:
          type: string
          format: uuid
        price:
          type: integer
        class:
          $ref: "#/components/schemas/ClassSummary"
    Invoice:
      type: object
      required: [id, userId, totalAmount, status, createdAt, paidAt, items]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        totalAmount:
          type: integer
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        createdAt:
          type: string
        paidAt:
          type: ["string", "null"]
        paymentUrl:
          type: ["string", "null"]
          description: URL halaman pembayaran Mayar. Diisi saat Mayar sudah diintegrasikan; null selama mode simulasi.
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
...